hv: Add ADMAC tracing helpers
Add an ADMAC tracing helper to m1n1. This lets us trace most of ADMAC
register accesses in an async mode while tracing written descriptors and
read transfer reports synchronously -- that's vital if we want to
monitor the content of buffers pointed at in descriptors.

We can't trace all accesses to ADMAC registers synchronously since that
would be too slow and audio is time-sensitive.

Signed-off-by: Martin Povišer <povik@cutebit.org>